Выносит общую логику обновления обработчиков и данных в класс Smart
